requirements.txt
Django==1.11.4
djangorestframework==3.6.4
drf-nested-routers==0.90.0
pkg-resources==0.0.0
psycopg2==2.7.3.1
pytz==2017.2
